Update asm references to be PIC.
This change is the result of a sed script and a little bit of hand tweaking and
causes references to globals in the assembly to go via the GOT. In order to
free up a register for the extra indirection, rbp is saved on the stack for
every function with a global reference.

This allows the code to be position independent. (i.e. to be compiled into a
shared object.)